mar01/requirements.txt
fastapi>=0.131.0,<0.132.0
uvicorn>=0.41.0,<0.42.0
sqlalchemy>=2.0.46,<3.0.0
asyncpg>=0.31.0,<0.32.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
email-validator>=2.2.0,<3.0.0
